# customer-service/db/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package sqlc

import (
	"database/sql"
)

type Customer struct {
	ID          int64
	UserID      sql.NullInt64
	Name        string
	Email       string
	CreditLimit string
}