# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PayLater API
  description: |
    PayLater API Gateway contract (port 8080).

    All protected routes require `Authorization: Bearer <JWT>` unless noted otherwise.

    Role rules:
    - `admin` — full administrative access where documented
    - `customer` — transaction/payback with customer ownership enforcement
    - `merchant` — merchant onboarding only; blocked from transaction/payback APIs

    Public `POST /auth/register` accepts `customer` and `merchant` only.
    `role=admin` self-registration returns 403.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local API Gateway
...
paths:
...
  /customers:
    get:
      tags: [Customers]
      summary: List customers
      description: Admin only.
      operationId: listCustomers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Customer list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Customers]
      summary: Create customer
      description: Admin only.
      operationId: createCustomer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
      responses:
        '201':
          description: Customer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /transactions:
    post:
      tags: [Transactions]
      summary: Create transaction
      description: |
        Admin or customer role.

        Customer role:
        - `customer_id` must belong to the authenticated user (403 otherwise)
        - credit limit enforced (400 when exceeded)

        Admin bypasses ownership and credit-limit enforcement.
      operationId: createTransaction
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionRequest'
      responses:
        '201':
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Validation error or credit limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Customer or merchant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /paybacks:
    post:
      tags: [Paybacks]
      summary: Create payback
      description: |
        Admin or customer role.

        Customer role:
        - `customer_id` must belong to the authenticated user (403 otherwise)
        - payback amount must be positive and not exceed remaining due (400)

        Admin bypasses ownership and remaining-due validation.
      operationId: createPayback
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaybackRequest'
      responses:
        '201':
          description: Payback created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Validation error or payback exceeds remaining due
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT obtained from POST /auth/login
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Role or ownership not authorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    Customer:
      type: object
      description: |
        GET /customers response item. Go serializes sqlc struct fields using PascalCase.
      properties:
        ID:
          type: integer
          format: int64
        Name:
          type: string
        Email:
          type: string
          format: email
        CreditLimit:
          type: string
          description: DECIMAL serialized as string
    CreateCustomerRequest:
      type: object
      required: [name, email, credit_limit]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        credit_limit:
          type: string
          description: DECIMAL amount as string
        user_id:
          type: integer
          format: int64
          nullable: true
...
    CreateTransactionRequest:
      type: object
      required: [customer_id, merchant_id, amount, commission, transaction_date]
      properties:
        customer_id:
          type: integer
          format: int64
        merchant_id:
          type: integer
          format: int64
        amount:
          type: string
        commission:
          type: string
        transaction_date:
          type: string
          format: date
          description: YYYY-MM-DD
    CreatePaybackRequest:
      type: object
      required: [customer_id, amount, payment_date]
      properties:
        customer_id:
          type: integer
          format: int64
        amount:
          type: string
        payment_date:
          type: string
          format: date
          description: YYYY-MM-DD
...